# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Monitoring platform 1000 Moustaches API",
    "description": "API documentation for the 1000 Moustaches monitoring platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/animals/": {
      "get": {
        "tags": [
          "Animals"
        ],
        "summary": "Get all animals",
        "description": "Retrieve a list of all animals in the system",
        "responses": {
          "200": {
            "description": "List of animals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Animal"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Animals"
        ],
        "summary": "Create new animal",
        "description": "Register a new animal in the system",
        "responses": {
          "201": {
            "description": "Animal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnimalRequest"
              }
            }
          }
        }
      }
    },
    "/animals/{id}": {
      "get": {
        "tags": [
          "Animals"
        ],
        "summary": "Get animal by ID",
        "description": "Retrieve a specific animal by its ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Animal details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Animals"
        ],
        "summary": "Update animal",
        "description": "Update an existing animal\\'s information",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Animal updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Animal"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnimalRequest"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Animals"
        ],
        "summary": "Delete animal",
        "description": "Remove an animal from the system",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Animal deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/animal-host-families/animal/{animalId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get host family history for animal",
        "description": "Retrieve all host family relationships for a specific animal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of host family relationships for the animal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/animal-host-families/hostFamily/{hostFamilyId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get animals hosted by host family",
        "description": "Retrieve all animals that have been hosted by a specific host family",
        "parameters": [
          {
            "name": "hostFamilyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Host family ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of animals hosted by the host family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/animal/{animalId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get host family history for animal",
        "description": "Retrieve all host family relationships for a specific animal",
        "parameters": [
          {
            "name": "animalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Animal ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of host family relationships for the animal",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/hostFamily/{hostFamilyId}": {
      "get": {
        "tags": [
          "Host Family Relations"
        ],
        "summary": "Get animals hosted by host family",
        "description": "Retrieve all animals that have been hosted by a specific host family",
        "parameters": [
          {
            "name": "hostFamilyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Host family ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of animals hosted by the host family",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalHostFamily"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Firebase authentication token"
      }
    },
    "schemas": {
      "Species": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Species name",
            "example": "Chat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "Animal": {
        "type": "object",
        "required": [
          "name",
          "icad"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Animal name",
            "example": "Whiskers"
          },
          "species": {
            "$ref": "#/components/schemas/Species"
          },
          "icad": {
            "type": "string",
            "description": "ICAD identification number",
            "example": "250268500123456"
          },
          "sexe": {
            "type": "string",
            "description": "Animal gender",
            "enum": [
              "M",
              "F"
            ],
            "example": "F"
          },
          "race": {
            "type": "string",
            "description": "Animal breed",
            "example": "European Shorthair"
          },
          "adopted": {
            "type": "boolean",
            "description": "Is adopted",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "HostFamily": {
        "type": "object",
        "required": [
          "name",
          "firstname"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Host family last name",
            "example": "Martin"
          },
          "firstname": {
            "type": "string",
            "description": "Host family first name",
            "example": "Marie"
          },
          "phone": {
            "type": "string",
            "description": "Phone number",
            "example": "+33123456789"
          },
          "mail": {
            "type": "string",
            "format": "email",
            "description": "Email address",
            "example": "marie.martin@example.com"
          },
          "address": {
            "type": "string",
            "description": "Address",
            "example": "456 Avenue des Fleurs, 69000 Lyon"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Is available for hosting",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
...
      "AnimalHostFamily": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier",
            "example": 1
          },
          "animal": {
            "$ref": "#/components/schemas/Animal"
          },
          "hostFamily": {
            "$ref": "#/components/schemas/HostFamily"
          },
          "entryDate": {
            "type": "string",
            "format": "date-time",
            "description": "Date when animal entered the host family",
            "example": "2023-12-15T10:30:00Z"
          }
        }
      },
...
      "CreateAnimalRequest": {
        "type": "object",
        "required": [
          "name",
          "icad"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Animal name",
            "example": "Whiskers"
          },
          "icad": {
            "type": "string",
            "description": "ICAD identification number",
            "example": "250268500123456"
          },
          "sexe": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ],
            "example": "F"
          },
          "race": {
            "type": "string",
            "example": "European Shorthair"
          }
        }
      },
      "UpdateAnimalRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Whiskers Updated"
          },
          "sexe": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ]
          },
          "race": {
            "type": "string"
          },
          "adopted": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Authorization": []
    }
  ]
}